+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - [x] 修复某些来自 `TooManyExpressionInspection` 的误报 / Fix some false positives from `TooManyExpressionInspection`
 - [x] 从脚本成员评估作用域上下文时，跳过表达式（通常是属性键）带参数的情况 / When evaluating the scope context from script members, skip cases where expressions (usually property keys) are parameterized
 - [x] 优化表达式带参数时，并集规则（unionConfig）和别名规则（aliasConfig）的匹配与展开的逻辑，从而避免意外的 OOM / Optimize the matching and expansion logic for union configs and alias configs when expressions are parameterized, avoiding unexpected OOM
+- [x] 匹配数据类型 `Parameter`（`$parameter`）时，要求必须形如标识符，并且不允许带参数 / When matching the data type `Parameter` (`$parameter`), it is required to be in the form of an identifier and is not allowed to be parameterized
 - [x] 其他优化与BUG修复 / Other optimizations and bug fixes
 
 ## 3.0.2 - 2026-09-08